[NES] Improved UNIF support and added emulation of a few UNIF-only boards
git-svn-id: svn://messdev.no-ip.org/mess@6153 9849079f-6229-0410-a07f-84406ce305ba

+++ src/mess/includes/nes.h
@@ -43,8 +43,8 @@ struct nes_struct
 	/* load-time cart variables which remain constant */
 	UINT8 trainer;
 	UINT8 battery;
-	UINT8 prg_chunks;
-	UINT8 chr_chunks;
+	UINT16 prg_chunks;	// a recently dumped multigame cart has 256 chunks of both PRG & CHR!
+	UINT16 chr_chunks;
 
 	UINT8 format;	// 1 = iNES, 2 = UNIF
